refactor(process-operational-projection): flatten directory structure
