+++ src/capabilities/dashboard/application/dashboard.operational-summary.readmodel.ts
@@ -1,4 +1,4 @@
-import type { ProcessAggregatedStatus } from '~/modules/process/features/operational-projection/application/operational-projection/operationalSemantics'
+import type { ProcessAggregatedStatus } from '~/modules/process/features/operational-projection/application/operationalSemantics'
 import type { TrackingActiveAlertReadModel } from '~/modules/tracking/features/alerts/application/projection/tracking.active-alert.readmodel'
 import {
   type TrackingOperationalAlertCategory,
