+++ src/modules/process/features/operational-projection/application/operationalSemantics.ts
