+++ src/modules/process/interface/http/process.http.mappers.ts
@@ -8,9 +8,9 @@ import type {
 } from '~/modules/process/application/process.records'
 import type { ProcessSyncSummaryReadModel } from '~/modules/process/application/usecases/list-processes-with-operational-summary.usecase'
 import type { ProcessEntity } from '~/modules/process/domain/process.entity'
-import { deriveProcessStatusFromContainers } from '~/modules/process/features/operational-projection/application/operational-projection/deriveProcessStatus'
-import { toOperationalStatus } from '~/modules/process/features/operational-projection/application/operational-projection/operationalSemantics'
-import type { ProcessOperationalSummary } from '~/modules/process/features/operational-projection/application/operational-projection/processOperationalSummary'
+import { deriveProcessStatusFromContainers } from '~/modules/process/features/operational-projection/application/deriveProcessStatus'
+import { toOperationalStatus } from '~/modules/process/features/operational-projection/application/operationalSemantics'
+import type { ProcessOperationalSummary } from '~/modules/process/features/operational-projection/application/processOperationalSummary'
 import type { CreateProcessInput } from '~/modules/process/interface/http/process.schemas'
 import {
   createTrackingOperationalSummaryFallback,
